requirements.txt
Flask
Flask-WTF
WTForms
Flask-SQLAlchemy
SQLAlchemy
pydantic
Flask-Login
flask-cors
psycopg2-binary
python-dotenv
